therapy_booking_app/requirements.txt
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-adk==0.1.0
requests==2.31.0
twilio==8.12.0
asyncio==3.4.3
datetime
logging
nltk==3.8.1
python-dateutil==2.8.2
